fix: unify word count to single canonical counter (courseWordCount.js)
Extract the inline pre-save word counter into server/src/utils/courseWordCount.js
as the single source of truth, and rewire every consumer to it:

- InteractiveCourse.js pre-save hook now delegates to countCourseWords()
- courseBuilder.js save/publish/validate paths drop the chars/5 estimator
- contentValidator.js section + total counts use the canonical counter
- validateCourses.js now reads sections (was walking modules and counting 0)

Delete five divergent/corrupted scripts (none imported): recalcWordCounts.js,
fixWordCounts.js, updateWordCounts.js, fixWordCountsFull.js, resetwc.js.

Behavior-preserving for the stored value (verbatim extraction); raises the
numbers validators/publish gate see, which is the intended fix.
